revisi

+++ database/migrations/2023_09_09_114518_create_settings_table.php
@@ -19,7 +19,7 @@ class CreateSettingsTable extends Migration
             $table->text('value');
             $table->timestamps();
             $table->datetime('created_by')->nullable();
-            $table->bigint('updated_by')->nullable();
+            $table->BigInteger('updated_by')->nullable();
         });
     }
